Integrate commitfest status
This commit adds a script to scrape the commitfest website, and mirror
some data in our database. This script is executed each day for open
commitfests, and once manually at the beginning for all.

Based on this info, we can link the commitfest patch pages and display
commitfest status even on the topic index.